Update dice_stimulator.py
updated code, the dice faces are stored as a list of strings in the DICE_FACES constant. The roll_dice() function selects the appropriate face based on the randomly generated number and prints it. The code will continue to roll the dice until the user enters something other than 'y'.

These improvements can enhance code readability, maintainability, and reusability.